requirements-test.txt
# Oracle AWR分析器 - 测试依赖配置
# {{CHENGQI: P1-TE-004 基础测试框架配置 - 2025-06-01 23:45:00 +08:00}}

# 首先安装基础依赖
-r requirements.txt

# 核心测试框架
pytest==7.4.4
pytest-django==4.8.0
pytest-cov==4.1.0
pytest-xdist==3.5.0
pytest-asyncio==0.23.2

# E2E测试
playwright==1.40.0
pytest-playwright==0.4.3

# 测试数据生成
factory-boy==3.3.0
faker==22.2.0

# 性能和基准测试
pytest-benchmark==4.0.0

# 测试报告和分析
pytest-html==4.1.1
pytest-json-report==1.5.0

# 数据库测试支持
pytest-mock==3.12.0

# 调试和开发工具
pytest-sugar==0.9.7
pytest-clarity==1.0.1

# 代码质量检查（开发时）
pylint==3.0.3
mypy==1.8.0
django-stubs==4.2.7

# 安全测试
bandit==1.7.5
safety==2.3.5 